Fixed a possible race condition in GenericObjectPool.Get()
GenericObjectPool.Get() first checks for the availability of a pool for the requested type. If there is one, it checks if the pool has an object available. When both conditions are met, an instance is retrieved from the pool and then returned.

Between the condition checking and retrieving the object another thread could 'steal' the last instance from the pool. This would throw an exception and not return null as specified.

By putting a lock around the condition checking and retrieving the instance we ensure this never happens.