Remove random datastructures and pools and crypto stuff (fixes #24)
Previously, there were sort-of-random datastructures, pools and crypto classes in this library. They ended up in here because I needed a place to put them for some project way back in the past. However, I didn't feel they belong in here. SwissArmyKnife is a library with truly generic and useful extension methods, classes, etc. These pools and structures were not generic and most implementations are incomplete, not thread safe or just not useful.